YukkiMusicBot: general refactoring

+++ YukkiMusic/utils/stream/autoclear.py
@@ -7,25 +7,17 @@
 #
 # All rights reserved.
 
+import contextlib
 import os
 
 from config import autoclean
 
 
 async def auto_clean(popped):
-    try:
+    with contextlib.suppress(Exception):
         rem = popped["file"]
         autoclean.remove(rem)
         count = autoclean.count(rem)
-        if count == 0:
-            if (
-                "vid_" not in rem
-                or "live_" not in rem
-                or "index_" not in rem
-            ):
-                try:
-                    os.remove(rem)
-                except:
-                    pass
-    except:
-        pass
+        if count == 0 and (("vid_" not in rem or "live_" not in rem or "index_" not in rem)):
+            with contextlib.suppress(Exception):
+                os.remove(rem)
